+++ common/src/main/java/com/iamkaf/amber/AmberMod.java
@@ -1,18 +1,14 @@
 package com.iamkaf.amber;
 
+import com.iamkaf.amber.api.commands.v1.SimpleCommands;
 import com.iamkaf.amber.api.core.v2.AmberModInfo;
 import com.iamkaf.amber.api.event.v1.events.common.CommandEvents;
-import com.iamkaf.amber.api.platform.v1.ModInfo;
-import com.iamkaf.amber.api.platform.v1.Platform;
 import com.iamkaf.amber.api.registry.v1.DeferredRegister;
 import com.iamkaf.amber.api.registry.v1.RegistrySupplier;
 import com.iamkaf.amber.platform.Services;
 import com.iamkaf.amber.util.Env;
 import com.iamkaf.amber.util.EnvExecutor;
-import com.mojang.brigadier.Command;
-import net.minecraft.commands.Commands;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.Item;
 
 import java.util.ArrayList;
@@ -42,22 +38,7 @@ public class AmberMod {
 
         CommandEvents.EVENT.register((dispatcher, registryAccess, environment) -> {
             Constants.LOG.info("Registering Amber commands for {}", Services.PLATFORM.getPlatformName());
-            dispatcher.register(Commands.literal("amber").executes(commandContext -> {
-                ModInfo amberInfo = Platform.getModInfo(Constants.MOD_ID);
-
-                // wat??
-                if (amberInfo == null) {
-                    commandContext.getSource().sendFailure(Component.literal("Amber mod info not found!"));
-                    return Command.SINGLE_SUCCESS;
-                }
-
-                commandContext.getSource()
-                        .sendSuccess(
-                                () -> Component.literal(amberInfo.name()).append(" - Version: " + amberInfo.version()),
-                                false
-                        );
-                return Command.SINGLE_SUCCESS;
-            }));
+            dispatcher.register(SimpleCommands.createBaseCommand(Constants.MOD_ID));
         });
     }
 }